perf(gui): optimize rendering and add unit tests
- Cache Visuals configuration to eliminate per-frame allocation
- Remove forced 10 FPS repaint loop
- Add frame state caching to reduce atomic operations
- Add unit tests for GUI utility functions and types
- Unify worker_count type to usize
- Minimize field visibility to private scope

+++ src/state.rs
@@ -316,8 +316,8 @@ impl AppState {
     }
 
     /// Returns the actual number of active worker threads.
-    pub fn get_actual_worker_count(&self) -> u64 {
-        self.worker_count.load(Ordering::Relaxed)
+    pub fn get_actual_worker_count(&self) -> usize {
+        self.worker_count.load(Ordering::Relaxed) as usize
     }
 
     pub fn set_actual_worker_count(&self, count: usize) {
